requirements.txt
aiohttp
discord.py @ git+https://github.com/Rapptz/discord.py@f6a74f74a7aed0879fc086805eae8873e745d0ea
python-dotenv
requests
urllib3
validators
pillow
python-dateutil
youtube_dl
PyNaCl